Extract oidc controller and service into independent module
