+++ packages/@n8n/backend-common/src/modules/__tests__/module-registry.test.ts
@@ -30,6 +30,7 @@ describe('eligibleModules', () => {
 			'source-control',
 			'dynamic-credentials',
 			'chat-hub',
+			'sso-oidc',
 		]);
 	});
 
@@ -46,6 +47,7 @@ describe('eligibleModules', () => {
 			'source-control',
 			'dynamic-credentials',
 			'chat-hub',
+			'sso-oidc',
 		]);
 	});
