+++ packages/@n8n/backend-common/src/modules/modules.config.ts
@@ -13,6 +13,7 @@ export const MODULE_NAMES = [
 	'source-control',
 	'dynamic-credentials',
 	'chat-hub',
+	'sso-oidc',
 ] as const;
 
 export type ModuleName = (typeof MODULE_NAMES)[number];
