+++ packages/cli/src/modules/sso-oidc/__tests__/oidc.service.ee.test.ts
@@ -9,16 +9,16 @@ import type { Cipher, InstanceSettings } from 'n8n-core';
 import * as client from 'openid-client';
 import { EnvHttpProxyAgent } from 'undici';
 
-import * as ssoHelpers from '../../sso-helpers';
-import { OIDC_PREFERENCES_DB_KEY } from '../constants';
-import { OidcService } from '../oidc.service.ee';
-
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { ForbiddenError } from '@/errors/response-errors/forbidden.error';
 import { type ProvisioningService } from '@/modules/provisioning.ee/provisioning.service.ee';
 import { Publisher } from '@/scaling/pubsub/publisher.service';
 import type { JwtService } from '@/services/jwt.service';
 import type { UrlService } from '@/services/url.service';
+import * as ssoHelpers from '@/sso.ee/sso-helpers';
+
+import { OIDC_PREFERENCES_DB_KEY } from '../constants';
+import { OidcService } from '../oidc.service.ee';
 
 jest.mock('undici', () => ({
 	// eslint-disable-next-line @typescript-eslint/naming-convention
