+++ packages/cli/src/modules/sso-oidc/constants.ts
